Move HttpContextExtensions.Redirect to its own file.
